requirements.txt
# Core dependencies
pyyaml>=6.0
tqdm>=4.64.0

# LLM providers (install based on your choice)
openai>=1.0.0
anthropic>=0.25.0

# Optional: for local models
# ollama>=0.1.0